ApiApplication/requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector==2.2.9
mysqlclient==2.0.3
PyMySQL==1.0.2
python-dotenv==0.19.0
redis==3.5.3
requests==2.26.0
SQLAlchemy==1.4.23
urllib3==1.26.6
Werkzeug==2.0.1
jsonpickle
